Setup unit tests for lambda paramter copying
