+++ hypo-test/hypo-test-data/src/scenario-13/java/scenario13/TestInterface.java
@@ -1,0 +1,6 @@
+package scenario13;
+
+public interface TestInterface<T1, T2> {
+
+    void apply(T1 t1, T2 t2);
+}
