+++ hypo-test/src/test/java/dev/denwav/hypo/test/scenarios/Scenario08Test.java
@@ -19,11 +19,9 @@
 package dev.denwav.hypo.test.scenarios;
 
 import dev.denwav.hypo.asm.hydrate.SuperConstructorHydrator;
-import dev.denwav.hypo.core.HypoConfig;
 import dev.denwav.hypo.hydrate.HydrationProvider;
 import dev.denwav.hypo.mappings.contributors.ChangeContributor;
 import dev.denwav.hypo.mappings.contributors.CopyMappingsDown;
-import dev.denwav.hypo.mappings.contributors.PropagateMappingsUp;
 import dev.denwav.hypo.test.framework.TestScenarioBase;
 import java.util.List;
 import org.cadixdev.lorenz.MappingSet;
@@ -31,67 +29,63 @@ import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-@DisplayName("[integration] Scenario 12 - Super Constructs from Inner Classes")
-public class Scenario12 extends TestScenarioBase {
+@DisplayName("[integration] Scenario 08 - Same-class Constructor Calls")
+public class Scenario08Test extends TestScenarioBase {
 
     @Override
     public @NotNull Env env() {
         return new Env() {
             @Override
             public @NotNull String forContext() {
-                return "scenario-12";
+                return "scenario-08";
             }
 
             @Override
             public @NotNull Iterable<HydrationProvider<?>> hydration() {
                 return List.of(SuperConstructorHydrator.create());
             }
-
-            @Override
-            public @NotNull HypoConfig.Builder config() {
-                return HypoConfig.builder().withParallelism(1);
-            }
         };
     }
 
     @Test
-    @DisplayName("Test super constructor param mappings from inner class")
-    void testSuperConstructorFromInnerClass() throws Exception {
+    @DisplayName("Test this() call parameter propagation")
+    void testThisPropagation() throws Exception {
         this.runTest(new TestCase() {
             @Override
             public @NotNull Iterable<ChangeContributor> changeContributors() {
-                return List.of(PropagateMappingsUp.create(), CopyMappingsDown.create());
+                return List.of(CopyMappingsDown.create());
             }
 
             @Override
             public @NotNull MappingSet startMappings() {
                 return parseTiny("""
-                    c    scenario12/BaseClass    scenario12/BaseClass
-                        m    (III)V    <init>    <init>
+                    c    scenario08/TestClass    scenario08/TestClass
+                        m    (II)V    <init>    <init>
                             p    1        i
                             p    2        j
-                            p    3        k
+                        m    (Ljava/lang/String;IJ)V    <init>    <init>
+                            p    1        someText
+                            p    2        i
+                            p    3        j
                     """);
             }
 
             @Override
             public @NotNull MappingSet finishMappings() {
                 return parseTiny("""
-                    c    scenario12/BaseClass    scenario12/BaseClass
-                        m    (III)V    <init>    <init>
+                    c    scenario08/TestClass    scenario08/TestClass
+                        m    (II)V    <init>    <init>
                             p    1        i
                             p    2        j
-                            p    3        k
-                    c    scenario12/TestClass$InnerClass    scenario12/TestClass$InnerClass
-                        m    (III)V    <init>    <init>
+                        m    (JJ)V    <init>    <init>
                             p    1        i
-                            p    2        j
-                            p    3        k
-                    c    scenario12/ChildClass    scenario12/ChildClass
-                        m    (III)V    <init>    <init>
-                            p    1        i
-                            p    2        j
-                            p    3        k
+                            p    3        j
+                        m    (Ljava/lang/String;IJ)V    <init>    <init>
+                            p    1        someText
+                            p    2        i
+                            p    3        j
+                        m    (Ljava/lang/String;Ljava/lang/String;)V    <init>    <init>
+                            p    1        someText
                     """);
             }
         });
